Вывод прогресса сделан в отдельном потоке

+++ lw1/Lelekov_Alexey/MonteCarloPi/MonteCarloPi/ThreadHandler.cpp
@@ -12,7 +12,7 @@ void ThreadHandler::Add(LPTHREAD_START_ROUTINE threadProc, LPVOID param)
 
 void ThreadHandler::JoinAll()
 {
-	WaitForMultipleObjects(threads.size(), threads.data(), true, INFINITE);
+	WaitForMultipleObjects((DWORD)threads.size(), threads.data(), true, INFINITE);
 }
 
 ThreadHandler::~ThreadHandler()
